+++ models/Stock.js
@@ -25,10 +25,6 @@ const StockSchema = new mongoose.Schema({
         type: String,
         required: true
     },
-    date: {
-        type: Date,
-        default: Date.now
-    },
 
 })
 module.exports = mongoose.model('Stock', StockSchema);
